Pagination to upload patterns(all)

+++ src/Component/Admin/Home/Tailor/Tailor.js
@@ -14,6 +14,10 @@ const Tailor = (props) => {
   const ref = useRef(null);
   // const [newTailorList, setNewTailorList] = useState(null);
   const [tailorList, setTailorList] = useState(null);
+  const [tailorCount, setTailorCount] = useState({
+    orderStatus: "",
+    count: 0
+  });
 
   $("button").on("click", function () {
     $("button").removeClass("selected");
@@ -21,21 +25,44 @@ const Tailor = (props) => {
   });
 
   useEffect(() => {
-    console.log("perorderuseeffect");
+    // console.log("perorderuseeffect");
     list = [];
     db.collection("TailorsDetails")
       .where("tailorStatus", "==", "Verified")
       .get()
       .then((data) => {
-        console.log("data", data);
+        // console.log("data", data);
         data.forEach((doc) => {
           list.push(doc.data());
         });
-
+        db.collection("TailorsDetails")
+          .where("tailorStatus", "==", "Registered")
+          .get()
+          .then((data) => {
+            setTailorCount({
+              orderStatus: "Registered",
+              count: data.size
+            });
+          });
         setTailorList(list);
         console.log("list", list);
       });
   }, []);
+
+  useEffect(() => {
+    if (tailorList !== null) {
+      // console.log("status", bookedCount.orderStatus, preorderList);
+      if (
+        tailorCount.tailorStatus === "Registered" &&
+        tailorList[0].tailorStatus === "Registered"
+      ) {
+        setTailorCount({
+          orderStatus: "Registered",
+          count: tailorList.length
+        });
+      }
+    }
+  }, [tailorList, tailorCount.tailorStatus]);
 
   const newTailorHandler = () => {
     console.log("verifiedorder");
@@ -88,36 +115,50 @@ const Tailor = (props) => {
   const acceptTailor = (tailor) => {
     console.log("Accepted");
     console.log(">>>>>>", tailor);
+    let value = window.confirm("Do you want Accept");
+    if (value) {
+      db.collection("TailorsDetails")
+        .doc(tailor.userId)
+        .update({
+          tailorStatus: "Verified"
+        })
+        .then(() => {
+          let data = [...tailorList];
 
-    db.collection("TailorsDetails")
-      .doc(tailor.userId)
-      .update({
-        tailorStatus: "Verified"
-      })
-      .then(() => {
-        let data = [...tailorList];
-
-        let filterdata = data.filter((d) => d.tailorId !== tailor.tailorId);
-        setTailorList(filterdata);
-      });
+          let filterdata = data.filter((d) => d.tailorId !== tailor.tailorId);
+          setTailorList(filterdata);
+        });
+    } else {
+      console.log("Not Accepted");
+    }
   };
   const rejectTailor = (tailor) => {
     console.log("Rejected");
-    db.collection("TailorsDetails")
-      .doc(tailor.userId)
-      .update({
-        tailorStatus: "Rejected"
-      })
-      .then(() => {
-        let data = [...tailorList];
+    let value = window.confirm("Do you want Reject");
+    if (value) {
+      db.collection("TailorsDetails")
+        .doc(tailor.userId)
+        .update({
+          tailorStatus: "Rejected"
+        })
+        .then(() => {
+          let data = [...tailorList];
 
-        let filterdata = data.filter((d) => d.tailorId !== tailor.tailorId);
-        setTailorList(filterdata);
-      });
+          let filterdata = data.filter((d) => d.tailorId !== tailor.tailorId);
+          setTailorList(filterdata);
+        });
+    } else {
+      console.log("nott rejected");
+    }
   };
 
   const deleteTailor = (tailor) => {
-    db.collection("TailorsDetails").doc(tailor.userId).delete();
+    let value = window.confirm("Do you want to Delete the tailor");
+    if (value) {
+      db.collection("TailorsDetails").doc(tailor.userId).delete();
+    } else {
+      console.log("Not Deleted");
+    }
   };
 
   let tailors = null;
@@ -126,17 +167,33 @@ const Tailor = (props) => {
   } else if (tailorList === "empty") {
     tailors = <h1>No Processing Orders</h1>;
   } else {
-    tailors = tailorList.map((tailor) => {
-      return (
-        <TailorView
-          tailor={tailor}
-          {...props}
-          acceptTailor={acceptTailor}
-          rejectTailor={rejectTailor}
-          deleteTailor={deleteTailor}
-        />
-      );
-    });
+    console.log("...>>>>>....", tailorList);
+
+    tailors = (
+      <>
+        {tailorList.tailorStatus === "Verified" ? (
+          <div class="box head">
+            <p class="id">Tailor Id</p>
+            <p class="tname">Name</p>
+            <p class="no">Number</p>
+            <p class="spec">Specalization</p>
+            <p class="add">Address</p>
+          </div>
+        ) : null}
+
+        {tailorList.map((tailor) => {
+          return (
+            <TailorView
+              tailor={tailor}
+              {...props}
+              acceptTailor={acceptTailor}
+              rejectTailor={rejectTailor}
+              deleteTailor={deleteTailor}
+            />
+          );
+        })}
+      </>
+    );
     // tailors = <TailorView {...props} />;
   }
 
@@ -152,9 +209,7 @@ const Tailor = (props) => {
         </button>
         <button type="button" onClick={() => newTailorHandler()}>
           New Tailor
-          <span className="new-count">
-            {tailorList !== null && tailorList.length}
-          </span>
+          <span className="new-count">{tailorCount.count}</span>
         </button>
         <button
           type="button"
@@ -170,7 +225,7 @@ const Tailor = (props) => {
           <LoadingBar color="#FF0000" ref={ref} />,
           document.getElementById("linear-loader")
         )}
-        <div id="content" className="content">
+        <div id="content" className="content m-0">
           {tailors}
         </div>
       </div>
